Prevent users from deleting other users contact

+++ backend/controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel");
 
 // Get all contacts
 // Route: GET /api/v1/contacts
-// Access: Public (for now)
+// Access: Private
 const getContacts = asyncHandler(async (req,res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 // Create contact
 // Route: GET /api/v1/contacts
-// Access: Public (for now)
+// Access: Private
 const viewContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
 
@@ -25,7 +25,7 @@ const viewContact = asyncHandler(async (req,res) => {
 
 // Create contact
 // Route: POST /api/v1/contacts
-// Access: Public (for now)
+// Access: Private
 const createContact = asyncHandler(async (req,res) => {
     console.log(req.body);
 
@@ -38,12 +38,18 @@ const createContact = asyncHandler(async (req,res) => {
         phone, 
         address,
         country,
-        state 
+        state
     } = req.body;
 
     if (!last_name || !first_name || !email || !phone || !address || !country || !state) {
         res.status(400);
         throw new Error("Provide details for required fields");
+    }
+
+    const emailExists = await Contact.findOne({ email });
+    if (emailExists) {
+        res.status(400)
+        throw new Error('Contact with that email already exists');
     }
 
     // If all required fields are satisfied
@@ -55,14 +61,15 @@ const createContact = asyncHandler(async (req,res) => {
         phone, 
         address,
         country,
-        state 
+        state,
+        user_id: req.user.id
     });
     res.status(201).json(contact);
 });
 
 // Create contact
 // Route: PUT /api/v1/contacts
-// Access: Public (for now)
+// Access: Private
 const updateContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
 
@@ -70,6 +77,11 @@ const updateContact = asyncHandler(async (req,res) => {
     if (!contact) {
         res.status(404);
         throw new Error('Contact not found');
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("You are not authorized to update this contact");
     }
 
     // Get the updated contact
@@ -83,13 +95,19 @@ const updateContact = asyncHandler(async (req,res) => {
 
 // Create contact
 // Route: DELETE /api/v1/contacts
-// Access: Public (for now)
+// Access: Private
 const deleteContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
       res.status(404);
       throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("You are not authorized to deelete this contact");
+    }
+
     await Contact.deleteOne({ _id: req.params.id });
     res.status(200).json(contact);
 });
